Feat: befor Repository

+++ backend/Listen_to_me/src/main/java/com/ssafy/a605/repository/BookmarkRepository.java
@@ -1,8 +1,11 @@
 package com.ssafy.a605.repository;
 
 import com.ssafy.a605.model.entity.Bookmark;
+import org.springframework.data.domain.Example;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
-
+public interface BookmarkRepository extends JpaRepository<Bookmark, Integer> {
+    
 }
